Turn fixture into plugin
